refactored to routers pattern with views and main app. fixes as per review. added pytest
